requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
openai>=1.30.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
